refactor: replace `LinkEntityModel` class
